Added JobAccountOperations (#2073)
* Added JobAccountOperations
* Added JobAccountOperations tests
* Disabled ServerDisasterRecoveryConfigurationOperations test because it fails after updating generated code from hydra specs.
NoContent was removed as a successful return code in hydra-specs-pr commit a2049b833c0e1bada9c5dc2e5fff0429a09aff77 , but azure-sdk-for-net was not re-generated from this spec and the test was not updated. I chatted with Frankie and we agreed to disable the test in order to unblock the code generation. I did not manually patch the generated code because that will just defer the failure to the next person to re-generated code from hydra specs.